backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
boto3>=1.34.0
httpx==0.26.0
pandas>=2.0.0
numpy>=1.26.0,<2.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
cdp-sdk>=0.0.1
web3>=6.0.0
